Dynamic growing season
it's not finished yet, but the comments have a plan for a dynamic growing season and the code that averages the temperatures of the months above 0. That's not what we want, but the solution for SPEI should be similar. I also switched out the SPEI file with the 12-month timescale with the 1-month one.