Dynamic GSD calculation
gsd now adjusts what months it averages based on temperature. Before it assumed the growing season was June-August for everywhere, but now it averages the SPEI for all months with an average temperature above 10C (50F). While this still isn't perfect, it accounts for warmer climates and places in the southern hemisphere.